requirements.txt
asyncio==3.4.3
beautifulsoup4==4.12.2
colorama==0.4.6
duckduckgo_search==4.1.1
md2pdf==1.0.1
playwright==1.40.0
openai~=1.6.1
python-dotenv~=1.0.0
pyyaml==6.0.1
selenium==4.15.2
webdriver-manager==4.0.1
uvicorn==0.24.0.post1
pydantic==2.5.1
fastapi==0.104.1
python-multipart==0.0.6
markdown==3.5.1
langchain==0.0.353
tavily-python==0.3.1
permchain==0.0.6
arxiv==2.0.0
PyMuPDF==1.23.6
requests==2.31.0
jinja2==3.1.2
aiofiles==23.2.1
mangum
